+++ packages/eslint-config-next/rules/graphql.js
@@ -1,0 +1,20 @@
+module.exports = {
+  overrides: [
+    {
+      files: ['*.graphql'],
+      extends: 'plugin:@graphql-eslint/operations-recommended',
+    },
+    {
+      files: ['**/*.{ts,tsx}'],
+      extends: ['plugin:@typescript-eslint/recommended'],
+      rules: {
+        'no-underscore-dangle': [
+          'error',
+          {
+            allow: ['__typename'],
+          },
+        ],
+      },
+    },
+  ],
+};