+++ packages/eslint-config-next/CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.3](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.3.2...@side/eslint-config-next@2.3.3) (2023-12-07)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [2.3.2](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.3.1...@side/eslint-config-next@2.3.2) (2023-12-07)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [2.3.1](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.3.0...@side/eslint-config-next@2.3.1) (2023-11-28)
+
+### Bug Fixes
+
+- **next:** allow underscore dangle for typename ([#583](https://github.com/reside-eng/lint-config/issues/583)) ([14408bc](https://github.com/reside-eng/lint-config/commit/14408bc76fba22eb1cdb166005a470c97db8d62b))
+
+# [2.3.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.2.0...@side/eslint-config-next@2.3.0) (2023-11-14)
+
+### Features
+
+- **next:** add support for codegen.ts ([#582](https://github.com/reside-eng/lint-config/issues/582)) ([40c4689](https://github.com/reside-eng/lint-config/commit/40c46899c309c6e6fc9ded55f5ac333e52760ebe))
+
+# [2.2.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.1.0...@side/eslint-config-next@2.2.0) (2023-10-24)
+
+### Features
+
+- update typescript config ([#562](https://github.com/reside-eng/lint-config/issues/562)) ([0a814d7](https://github.com/reside-eng/lint-config/commit/0a814d7ac0513db129556aa2cf206f241955e855))
+
+# [2.1.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@2.0.0...@side/eslint-config-next@2.1.0) (2023-06-23)
+
+### Features
+
+- **eslint-config-next:** add graphql operations recommended rules ([#550](https://github.com/reside-eng/lint-config/issues/550)) ([993615d](https://github.com/reside-eng/lint-config/commit/993615d07ad123d8ae55c03ca5e77aa4b0ce34ef))
+
+# [2.0.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.1.0...@side/eslint-config-next@2.0.0) (2023-03-29)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+# [2.0.0-alpha.1](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.1.0...@side/eslint-config-next@2.0.0-alpha.1) (2023-03-29)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+# [1.1.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.21...@side/eslint-config-next@1.1.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#492](https://github.com/reside-eng/lint-config/issues/492)) ([54c907b](https://github.com/reside-eng/lint-config/commit/54c907bc7d6dee491fbac3ecb9769284554785cf)), closes [#488](https://github.com/reside-eng/lint-config/issues/488)
+
+# [1.1.0-alpha.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.21...@side/eslint-config-next@1.1.0-alpha.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#488](https://github.com/reside-eng/lint-config/issues/488)) ([6b10c5e](https://github.com/reside-eng/lint-config/commit/6b10c5ea54a9e68f7e3b04499a48a85b704a93cc))
+
+## [1.0.21](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.20...@side/eslint-config-next@1.0.21) (2023-03-13)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#468](https://github.com/reside-eng/lint-config/issues/468)) ([2ab68cd](https://github.com/reside-eng/lint-config/commit/2ab68cd8a2e8cc06899a50beccd99f2b3c9e8d7a))
+
+## [1.0.20](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.19...@side/eslint-config-next@1.0.20) (2023-02-27)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [1.0.19](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.18...@side/eslint-config-next@1.0.19) (2023-02-27)
+
+### Bug Fixes
+
+- **deps:** update dependency @next/eslint-plugin-next to v13 ([#428](https://github.com/reside-eng/lint-config/issues/428)) ([cd90736](https://github.com/reside-eng/lint-config/commit/cd90736fb181b1dea0ff3f9270824af8706b174a))
+
+## [1.0.18](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.17...@side/eslint-config-next@1.0.18) (2023-02-27)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [1.0.17](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.16...@side/eslint-config-next@1.0.17) (2022-10-12)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [1.0.16](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.15...@side/eslint-config-next@1.0.16) (2022-10-11)
+
+**Note:** Version bump only for package @side/eslint-config-next
+
+## [1.0.15](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.14...@side/eslint-config-next@1.0.15) (2022-10-03)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#415](https://github.com/reside-eng/lint-config/issues/415)) ([2e6788f](https://github.com/reside-eng/lint-config/commit/2e6788f1d1a99b018345544bdef690a3faf13170))
+
+## [1.0.14](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.13...@side/eslint-config-next@1.0.14) (2022-09-19)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#411](https://github.com/reside-eng/lint-config/issues/411)) ([2b1935a](https://github.com/reside-eng/lint-config/commit/2b1935a8f6a5a0f70df179d30b70d75c3f6fc6d7))
+
 ## [1.0.13](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-next@1.0.12...@side/eslint-config-next@1.0.13) (2022-08-15)
 
 ### Bug Fixes